// src/gs-api/src/models/mvt-stk-dto.ts
/* tslint:disable */
import { ArticleDto } from './article-dto';
export interface MvtStkDto {
  id?: number;
  dateMvt?: string;
  quantite?: number;
  article?: ArticleDto;
  typeMvt?: 'ENTREE' | 'SORTIE' | 'CORRECTION_POS' | 'CORRECTION_NEG';
  sourceMvt?: 'COMMANDE_CLIENT' | 'COMMANDE_FOURNISSEUR' | 'VENTE';
  idEntreprise?: number;
}